--- extras/samplePrograms/FirstProgram.osh
@@ -1,3 +1,0 @@
-int a = 1;
-int b = 2;
-exit(0);
