--- extras/samplePrograms/example.pblasm
@@ -1,1 +1,0 @@
-HLT
