--- extras/samplePrograms/try-run-all-instructions.sasm
@@ -1,128 +1,0 @@
-AAA
-AAD
-AAM
-AAS
-ADC
-ADD
-AND
-CALL
-CBW
-CLC
-CLD
-CLI
-CMC
-CMP
-CMPSB
-CMPSW
-CWD
-DAA
-DAS
-DEC
-DIV
-HLT
-IDIV
-IMUL
-IN
-INC
-INT
-INTO
-IRET
-JA
-JAE
-JB
-JBE
-JC
-JCXZ
-JE
-JG
-JGE
-JL
-JLE
-JMP
-JNA
-JNAE
-JNB
-JNBE
-JNC
-JNE
-JNG
-JNGE
-JNL
-JNLE
-JNO
-JNP
-JNS
-JNZ
-JO
-JP
-JPE
-JPO
-JS
-JZ
-LAHF
-LDS
-LEA
-LES
-LODSB
-LODSW
-LOOP
-LOOPE
-LOOPNE
-LOOPNZ
-LOOPZ
-MOVSB
-MOVSW
-MUL
-NEG
-NOP
-NOT
-OR
-OUT
-POP
-POPA
-POPF
-PUSH
-PUSHA
-PUSHF
-RCL
-RCR
-REP
-REPE
-REPNE
-REPNZ
-REPZ
-RET
-RETF
-ROL
-ROR
-SAHF
-SAL
-SAR
-SBB
-SCASB
-SCASW
-SHL
-SHR
-STC
-STD
-STI
-STOSB
-STOSW
-SUB
-TEST
-XCHG
-XLATB
-XOR
-; MOV AX, 10
-; MOV BX, 20
-; MOV CX, 30
-; MOV DX, 40
-; MOV [3], 10
-; MOV [2 3], 20
-; MOV [1 2 3], 30
-; MOV [1 2 3 4], 40 
-; MOV CX, AX
-; MOV DX, BX
-; MOV [2], AX
-; MOV [3], [2]
-; MOV BX, [2]
