+++ gbvm/SASM/src/tmp.c
@@ -1,5 +1,5 @@
-#include "sasm_opcodes.h"
-#include "sasm_flags.h"
+#include "tmp.h"
+#include "sasm_memory.h"
 
 Error __AAA(CPU* cpu, Memory* mem, Word* operand1, Word* operand2)
 {
@@ -369,279 +369,5 @@ Error __XLATB(CPU* cpu, Memory* mem, Word* operand1, Word* operand2)
 Error __XOR(CPU* cpu, Memory* mem, Word* operand1, Word* operand2)
 {
     printf("CALLED __XOR\n");
-    return ERR_OK;
-}
-
-Error __DONOP(CPU* cpu, Memory* mem, Word* operand1, Word* operand2)
-{
-    printf("\nDO No OPeration\n");
-    return ERR_OK;
-}
-
-Error __CLRCF(CPU* cpu, Memory* mem, Word* operand1, Word* operand2)
-{
-    printf("\nBefore : %d", getFlag(CARRY, cpu));
-    setFlag(CARRY, cpu, false);
-    printf("\nAfter  : %d\n", getFlag(CARRY, cpu));
-    return ERR_OK;
-}
-
-Error __TGLCF(CPU* cpu, Memory* mem, Word* operand1, Word* operand2)
-{
-    printf("\nBefore : %d", getFlag(CARRY, cpu));
-    setFlag(CARRY, cpu, !getFlag(CARRY, cpu));
-    printf("\nAfter  : %d\n", getFlag(CARRY, cpu));
-    return ERR_OK;
-}
-
-Error __SHUTS(CPU* cpu, Memory* mem, Word* operand1, Word* operand2)
-{
-    printf("\nSHUTing down System\n");
-    setFlag(HALT, cpu, true);
-    return ERR_OK;
-}
-
-Error __DECR(CPU* cpu, Memory* mem, Word* operand1, Word* operand2)
-{
-    printf("\nBefore : %d", *operand1);
-
-    bool overflow = *operand1 == (-MAX_WORD);
-    bool auxiliary = (((*operand1 & 0x0F) + (-1 & 0x0f)) > 0x0f);
-    *operand1 = *operand1 + -1;
-
-    setFlag(OVERFLOW, cpu, overflow);
-    setFlag(ZERO, cpu, (*operand1 == 0));
-    setFlag(SIGN, cpu, (*operand1 < 0));
-    setFlag(AUX, cpu, auxiliary);
-    checkAndSetParity(cpu, *operand1);
-
-    printf("\nAfter  : %d\n", *operand1);
-    return ERR_OK;
-}
-
-Error __NEG(CPU* cpu, Memory* mem, Word* operand1, Word* operand2)
-{
-    printf("CALLED __NEG\n");
-    return ERR_OK;
-}
-
-Error __GOTO(CPU* cpu, Memory* mem, Word* operand1, Word* operand2)
-{
-    printf("\nUnconditional Jump to %d\n", *operand1);
-    cpu->registers.IP = *operand1 - 1;
-    return ERR_OK;
-}
-
-Error __JA(CPU* cpu, Memory* mem, Word* operand1, Word* operand2)
-{
-    if (!getFlag(CARRY, cpu) && !getFlag(ZERO, cpu)) {
-        printf("\nJUMPING TO %d\nREASON : __JA / __JNBE \n", *operand1);
-        cpu->registers.IP = *operand1 - 1;
-    }
-    return ERR_OK;
-}
-
-Error __JNC(CPU* cpu, Memory* mem, Word* operand1, Word* operand2)
-{
-    if (!getFlag(CARRY, cpu)) {
-        printf("\nJUMPING TO %d\nREASON : __JAE / __JNB / __JNC\n", *operand1);
-        cpu->registers.IP = *operand1 - 1;
-    }
-    return ERR_OK;
-}
-
-Error __JC(CPU* cpu, Memory* mem, Word* operand1, Word* operand2)
-{
-    if (getFlag(CARRY, cpu)) {
-        printf("\nJUMPING TO %d\nREASON : __JB /__JNAE / __JC\n", *operand1);
-        cpu->registers.IP = *operand1 - 1;
-    }
-    return ERR_OK;
-}
-
-Error __JBE(CPU* cpu, Memory* mem, Word* operand1, Word* operand2)
-{
-    if (getFlag(CARRY, cpu) || getFlag(ZERO, cpu)) {
-        printf("\nJUMPING TO %d\nREASON : __JBE / __JNA\n", *operand1);
-        cpu->registers.IP = *operand1 - 1;
-    }
-    return ERR_OK;
-}
-
-Error __JCXZ(CPU* cpu, Memory* mem, Word* operand1, Word* operand2)
-{
-    if (cpu->registers.CX.full == 0) {
-        printf("\nJUMPING TO %d\nREASON : __JCXZ\n", *operand1);
-        cpu->registers.IP = *operand1 - 1;
-    }
-    return ERR_OK;
-}
-
-Error __JZ(CPU* cpu, Memory* mem, Word* operand1, Word* operand2)
-{
-    if (getFlag(ZERO, cpu)) {
-        printf("\nJUMPING TO %d\nREASON : __JE / __JZ\n", *operand1);
-        cpu->registers.IP = *operand1 - 1;
-    }
-    return ERR_OK;
-}
-
-Error __JG(CPU* cpu, Memory* mem, Word* operand1, Word* operand2)
-{
-    if (getFlag(SIGN, cpu) == getFlag(OVERFLOW, cpu) && !getFlag(ZERO, cpu)) {
-        printf("\nJUMPING TO %d\nREASON : __JG / __JNLE\n", *operand1);
-        cpu->registers.IP = *operand1 - 1;
-    }
-    return ERR_OK;
-}
-
-Error __JGE(CPU* cpu, Memory* mem, Word* operand1, Word* operand2)
-{
-    if (getFlag(SIGN, cpu) == getFlag(OVERFLOW, cpu)) {
-        printf("\nJUMPING TO %d\nREASON : __JGE / __JNL\n", *operand1);
-        cpu->registers.IP = *operand1 - 1;
-    }
-    return ERR_OK;
-}
-
-Error __JL(CPU* cpu, Memory* mem, Word* operand1, Word* operand2)
-{
-    if (getFlag(SIGN, cpu) != getFlag(OVERFLOW, cpu)) {
-        printf("\nJUMPING TO %d\nREASON : __JL / __JNGE\n", *operand1);
-        cpu->registers.IP = *operand1 - 1;
-    }
-    return ERR_OK;
-}
-
-Error __JLE(CPU* cpu, Memory* mem, Word* operand1, Word* operand2)
-{
-    if (getFlag(SIGN, cpu) != getFlag(OVERFLOW, cpu) || getFlag(ZERO, cpu)) {
-        printf("\nJUMPING TO %d\nREASON : __JLE / __JNG\n", *operand1);
-        cpu->registers.IP = *operand1 - 1;
-    }
-    return ERR_OK;
-}
-
-Error __JNZ(CPU* cpu, Memory* mem, Word* operand1, Word* operand2)
-{
-    if (!getFlag(ZERO, cpu)) {
-        printf("\nJUMPING TO %d\nREASON : __JNE / __JNZ\n", *operand1);
-        cpu->registers.IP = *operand1 - 1;
-    }
-    return ERR_OK;
-}
-
-Error __JNO(CPU* cpu, Memory* mem, Word* operand1, Word* operand2)
-{
-    if (!getFlag(OVERFLOW, cpu)) {
-        printf("\nJUMPING TO %d\nREASON : __JNO\n", *operand1);
-        cpu->registers.IP = *operand1 - 1;
-    }
-    return ERR_OK;
-}
-
-Error __JNP(CPU* cpu, Memory* mem, Word* operand1, Word* operand2)
-{
-    if (!getFlag(PARITY, cpu)) {
-        printf("\nJUMPING TO %d\nREASON : __JNP / __JPO\n", *operand1);
-        cpu->registers.IP = *operand1 - 1;
-    }
-    return ERR_OK;
-}
-
-Error __JNS(CPU* cpu, Memory* mem, Word* operand1, Word* operand2)
-{
-    if (!getFlag(SIGN, cpu)) {
-        printf("\nJUMPING TO %d\nREASON : __JNS\n", *operand1);
-        cpu->registers.IP = *operand1 - 1;
-    }
-    return ERR_OK;
-}
-
-Error __JO(CPU* cpu, Memory* mem, Word* operand1, Word* operand2)
-{
-    if (getFlag(OVERFLOW, cpu)) {
-        printf("\nJUMPING TO %d\nREASON : __JO\n", *operand1);
-        cpu->registers.IP = *operand1 - 1;
-    }
-    return ERR_OK;
-}
-
-Error __JP(CPU* cpu, Memory* mem, Word* operand1, Word* operand2)
-{
-    if (getFlag(PARITY, cpu)) {
-        printf("\nJUMPING TO %d\nREASON : __JP / __JPE\n", *operand1);
-        cpu->registers.IP = *operand1 - 1;
-    }
-    return ERR_OK;
-}
-
-Error __JS(CPU* cpu, Memory* mem, Word* operand1, Word* operand2)
-{
-    if (getFlag(SIGN, cpu)) {
-        printf("\nJUMPING TO %d\nREASON : __JS\n", *operand1);
-        cpu->registers.IP = *operand1 - 1;
-    }
-    return ERR_OK;
-}
-
-Error __CPY(CPU* cpu, Memory* mem, Word* operand1, Word* operand2)
-{
-    printf("\nBefore : %d, %d", *operand1, *operand2);
-    *operand1 = *operand2;
-    printf("\nAfter  : %d\n", *operand1);
-    return ERR_OK;
-}
-
-Error __ADC(CPU* cpu, Memory* mem, Word* operand1, Word* operand2)
-{
-    printf("\nBefore : %d, %d", *operand1, *operand2);
-
-    bool overflow = (*operand2 > 0 && *operand1 > (MAX_WORD - *operand2));
-    bool auxiliary = (((*operand1 & 0x0F) + (*operand2 & 0x0f)) > 0x0f);
-    *operand1 = *operand1 + *operand2;
-    *operand1 = *operand1 + (getFlag(CARRY, cpu) ? 1 : 0);
-
-    setFlag(CARRY, cpu, overflow);
-    setFlag(OVERFLOW, cpu, overflow);
-    setFlag(ZERO, cpu, (*operand1 == 0));
-    setFlag(SIGN, cpu, (*operand1 < 0));
-    setFlag(AUX, cpu, auxiliary);
-    checkAndSetParity(cpu, *operand1);
-
-    printf("\nAfter  : %d\n", *operand1);
-    return ERR_OK;
-}
-
-Error __ADD(CPU* cpu, Memory* mem, Word* operand1, Word* operand2)
-{
-    printf("\nBefore : %d, %d", *operand1, *operand2);
-
-    bool overflow = (*operand2 > 0 && *operand1 > (MAX_WORD - *operand2));
-    bool auxiliary = (((*operand1 & 0x0F) + (*operand2 & 0x0f)) > 0x0f);
-    *operand1 = *operand1 + *operand2;
-
-    setFlag(CARRY, cpu, overflow);
-    setFlag(OVERFLOW, cpu, overflow);
-    setFlag(ZERO, cpu, (*operand1 == 0));
-    setFlag(SIGN, cpu, (*operand1 < 0));
-    setFlag(AUX, cpu, auxiliary);
-    checkAndSetParity(cpu, *operand1);
-
-    printf("\nAfter  : %d\n", *operand1);
-    return ERR_OK;
-}
-
-Error __AND(CPU* cpu, Memory* mem, Word* operand1, Word* operand2)
-{
-    printf("\nBefore : %d, %d", *operand1, *operand2);
-
-    *operand1 = *operand1 & *operand2;
-
-    setFlag(ZERO, cpu, (*operand1 == 0));
-    setFlag(SIGN, cpu, (*operand1 < 0));
-    checkAndSetParity(cpu, *operand1);
-
-    printf("\nAfter  : %d\n", *operand1);
     return ERR_OK;
 }
